requirements.txt
google-generativeai==0.3.2
ipython==8.21.0